+++ src/pages/backstage.tsx
@@ -4,7 +4,7 @@ import { graphql, Link } from 'gatsby';
 
 import Layout from '../components/layout';
 import Animation from '../components/animation';
-import ContactCTA from '../components/contact-cta';
+import ContactCTA from '../components/cta-contact';
 
 import {
   pageWrap,
