+++ src/pages/consulting.tsx
@@ -6,7 +6,7 @@ import Layout from '../components/layout';
 import BlogPreview from '../components/blog-preview';
 import { Tabs, Tab } from '../components/tabs/tabs';
 import Animation from '../components/animation';
-import ContactCTA from '../components/contact-cta';
+import ContactCTA from '../components/cta-contact';
 
 import {
   pageWrap,
